Conflicts:
	Applications/MPExtended.Applications.WebMediaPortal/Controllers/StreamController.cs

+++ Services/MPExtended.Services.StreamingService/Units/FFMpegLogParsingUnit.cs
@@ -36,9 +36,9 @@ namespace MPExtended.Services.StreamingService.Units
         public bool LogProgress { get; set; }
         private Reference<WebTranscodingInfo> data;
         private Thread processThread;
-        private int startPosition;
+        private long startPosition;
 
-        public FFMpegLogParsingUnit(Reference<WebTranscodingInfo> save, int startPosition) 
+        public FFMpegLogParsingUnit(Reference<WebTranscodingInfo> save, long startPosition) 
         {
             data = save;
             this.startPosition = startPosition;
@@ -76,7 +76,7 @@ namespace MPExtended.Services.StreamingService.Units
             return true;
         }
 
-        private static void ParseOutputStream(Stream outputStream, Reference<WebTranscodingInfo> saveData, int startPosition, bool logMessages, bool logProgress)
+        private static void ParseOutputStream(Stream outputStream, Reference<WebTranscodingInfo> saveData, long startPosition, bool logMessages, bool logProgress)
         {
             StreamReader reader = new StreamReader(outputStream);
